Store SegmentIndex as int32 instead of uint8 (all code setting the segment index works with int32 and there are no checks on downcasts which can cause overflow with 256+ segments)
#rb Jason.Nadro

[CL 35767222 by kenzo terelst in ue5-main branch]

+++ Engine/Plugins/FX/Niagara/Source/Niagara/Private/NiagaraRendererMeshes.cpp
@@ -1761,7 +1761,7 @@ void FNiagaraRendererMeshes::GetDynamicRayTracingInstances(FRayTracingMaterialGa
 				bIsInstancedStereo,
 				false // bNeedsPrevTransform
 			);
-			MeshBatch.SegmentIndex = uint8(SectionIndex);
+			MeshBatch.SegmentIndex = SectionIndex;
 			MeshBatch.LODIndex = uint8(LODModel.LODIndex);
 
 #if !(UE_BUILD_SHIPPING || UE_BUILD_TEST)
